requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.3.0
httpx==0.27.0
python-dotenv==1.0.1

# Gemini
google-generativeai==0.7.2
langchain-google-genai==1.0.6
langchain-core==0.2.1

# MCP
mcp==0.1.0
